+++ buildSrc/src/main/kotlin/Constants.kt
@@ -33,9 +33,8 @@ object Constants {
     }
 
     object Android {
-        const val compileSdkVersion = 35
-        const val targetSdkVersion = 35
-        const val minSdkVersion = 21
-        const val buildToolsVersion = "34.0.0"
+        const val compileSdkVersion = 36
+        const val targetSdkVersion = 36
+        const val minSdkVersion = 23
     }
 }